requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
urllib3==2.4.0
pyTelegramBotAPI==4.27.0
python-dotenv==1.0.1